Update patch set 1
Patch Set 1:

> I wonder - would it be possible to fire both in a transition
 > period, before we are reasonably sure people have upgraded to the
 > latest richdocuments?

Yes, I was thinking about the same as well. This can be done but will complicate the logic a bit on the WOPI host side. Especially for those WOPI implementations who wants to make sure that their implementation work for both older and newer online versions; in order to achieve that, they will listen to both 'rev-history' and 'UI_FileVersions' message. They have to make sure that they handle this case well - when both are fired. We can manage it in oc and nc ourself, but this needs to be made clear to other WOPI hosts. If it sounds okay to you, I can modify the patch to fire both.

Patch-set: 1
Label: Verified=0