Update patch set 1
Patch Set 1:

Overall looks good to me. I noticed one problem. When a document with a long comment at the bottom of the page is opened, the document bounds are adjusted correctly and user is able to scroll to see the comment.

But when user presses the enter key just before the comment such that a new page is created and then presses backspace such that new page is removed from the document, the document bounds get confused and you get a large grey area below the actual document. Sure, it's probably not a big blocker but just mentioned as I noticed it along the way.

Patch-set: 1
Reviewer: Gerrit User 1000897 <1000897@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0