Update patch set 1
Patch Set 1:

A quick grep suggests that when the disk becomes full, we set the readonly flag. But I wouldn't simulate the diskfull condition to check this. I would just open the browser console and do a map.setPermission(readonly), which should be enough to take care of this condition.

Patch-set: 1