Update patch set 1
Patch Set 1:

Sorry for the delay.

So, what if we reduce the width of the document-name-input to 140px and make padding-right to 20px on hover. That way we can put the document-name-input.{hide|show}() thing in mobilify and unmobilify and get rid of the 860 fileshowname limit. 

I tried it on local and it looks good, I think. One can argue that big file names would be a problem but that would be the case always in its current position when the screen size is low. Can you please try that?

Patch-set: 1