wsd: remove markToDestroy
Not really needed because we should always
stop the DocBroker when we are left with no sessions.

As for removing the last session, which is used to
store the document after saving (i.e. upload to storage),
we detect that by checking if the session isCloseFrame.
This should be reliable because in this scenario (and
only one related to markToDestroy) the trigger for
the removal (of last session), saving, and stopping
are all done because the client disconnected. Hence
the expectation that isCloseFrame should be true.

Change-Id: I762181084dc1275ad9b0b3cb16ad3426535c401d