Update patch set 2
Patch Set 2: Code-Review+2 Verified+1

Okay; looks good to me. 

As an additional feature, I think it's better if we attach a tooltip for the document name, so that if the document name doesn't fit in the box, atleast hovering over the box shows the full name of the document. But that can be done in separate commit.

Thanks for fixing this.

Patch-set: 2
Reviewer: Gerrit User 1000897 <1000897@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2