Update patch set 1
Patch Set 1:

Hi Emre,

Patch looks great. But it seems to have caught by our linting tool. Did you try `make` inside loleaflet/ and did it compile for you? There are some minor indentation fixes that needs to be made (tab, space mixing, etc.)

Other than that, I tried your patch and when a new sheet is added by this button, it should get selected after being added as well which is not the case with this patch. Can you please take care of this as well?

Patch-set: 1
Label: Verified=0