Update patch set 2
Patch Set 2:

(3 comments)

A good start; I had hoped to have the soft limit as an overlay that sits over the users' UI in way that makes it very visible, but allows them to continue to use the suite while remaining aware that this is "unsupported".

I guess we also want to check that if the limit presentation is disabled or high - that the product name does not contain 'libreoffice' for TM reasons - and that there is some suitable warning, compile-failure etc. in this case - but happy to add that.

Patch-set: 2
Label: Verified=0