Update patch set 2
Patch Set 2:

So, I looked at the plus button svg for insertsheet in core and there isn't any as such - we draw the plus button manually in core.git. So your best option is to use the 'plus.svg' in online.git which is also used for zoomin thing in our toolbar. Can you rename the zoomin label for the plus.svg in toolbar.css and use it for zoomin as well as for this new awesome insertsheet feature?

Patch-set: 2