Revert all related loleaflet IME changes
... that can potentially affect the html/css layout. We have a problem
in the front-end where map gets incorrectly positioned due to the moving
candidate window. With this revert, the candidate window is thrown back
at where it was (to the top-left) and the problem we are facing (as
mentioned above) dissappears as well. Let's fix that problem first
before re-revert it.

This reverts commit 98b53ae956782ce7366a2d4b29e80ce31b748058.

Change-Id: I2882900246efc068e514de134dc3fdd07f1ef024