Don't show the ruler if the window has too small height
Of course, for this to really be worthy of being called "responsive"
web design, we should update this decision also if the window is
re-sized dynamically, like when a desktop browser's window is
arbitrarily re-sized by the user, or a mobile device is rotated
between portrait and landscape orientation, or a mobile device enters
or leaves some kind of split-screen mode where another application
than the browser shares the display.

Also, it would be nice if the window height limit used here, 1280, was
some kind of global variable. (For width limits used to detect mobile
devices, we use 768 or 768px here and there in our JS code. (And,
interestingly, 767px, too.) Ideally this 1280 and those 768s should be
defined in some common place. No idea how to do that correctly in
JavaScript.

Change-Id: I02bb54f309bc116ae1a0b5264e72b9ba674f56e4