Don't show the ruler if the window is small
Of course, for this to really be worthy of being called "responsive"
web design, we should update this decision also if the window is
re-sized dynamically, like when a desktop browser's window is
arbitrarily re-sized by the user, or a mobile device is rotated
between portrait and landscape orientation, or a mobile device enters
or leaves some kind of split-screen mode.

Also, it would be nice if the magic window width limit used here, 768,
was some kind of global variable. We use 768 or 768px here and there
in our JS code. (And, interestingly, 767px, too.)
loleaflet/dist/toolbar/toolbar.js has a var mobileWidth = 768, would
that be usable elsewhere, too? And does it have to be a "variable",
does JavaScript have constants? Asking for a friend.

Change-Id: I02bb54f309bc116ae1a0b5264e72b9ba674f56e4