Don't show the ruler if the window has too small height
Of course, for this to really be worthy of being called "responsive"
web design, we should update this decision also if the window is
re-sized dynamically, like when a desktop browser's window is
arbitrarily re-sized by the user, or a mobile device is rotated
between portrait and landscape orientation, or a mobile device enters
or leaves some kind of split-screen mode where another application
than the browser shares the display.

We use 5% of the window height as the limit for how much vertical
space the ruler can occupy. Sadly I don't know if there is any way to
get the height attribute of .loleaflet-ruler in
loleaflet/dist/leaflet.css (which is 20px) in Map.js, so I hard-code
the limit of window height as the magic value 400 (5% of 400 is 20).

Change-Id: I02bb54f309bc116ae1a0b5264e72b9ba674f56e4