Avoid rendering / sending the same tile twice
When we're doing the prerendering we also need to register
a tileBeingRendered object, so we know that the given tile
will arrive soon. In earlier version of the code, rendering
and sending of the tile was not separated in time, but now it is,
so we need create a tileBeingRendered object even if we don't
subscribe the client to it yet.

(cherry picked from commit 5260eae05f89056d4ffc1bdfad1927c3c5f9706d)

Change-Id: I1374c22e5ffebe1d6fcc6e37261ddcedeb9066ec