Fix tilesBeingRendered tracking by client
I changed the code in this commit:
c2a5f6acb0f1e93f19104b761661c852d930fb9e

To make kit send a tilecombine message even if it does not
send actual tile data so we can track that the rendering was
done and so we can update the clients' _tilesBeingRendered
list. The issue is that tileBeingRendered object
belongs to not only one client, but more and so we don't
know which client gets the actual empty tile response.

So revert this method and rather use a smaller timeout for "waiting" the
arrival of the rendered tile.

(cherry picked from commit 07e99ad336c87f3d7e211abaaf18639be27c2929)

Change-Id: I2dbbab1a62b81cbbb5314f2f37fdbc3415c69130