Better to have a smaller tiles-on-fly limit
It's good if this limit big enough to send all the tiles of the
current visible area at once, so the tiles arrive at the same
time to the client by zoom or scroll. Now this value is set to a bit
bigger so if we have a small amount of tiles before zoom / scroll we'll
still see the same on the client side.

(cherry picked from commit a1e2acd5effb1c805f06c78040b71d30fb821be4)

Change-Id: I8e28dbf6197fe2f683fe9528e9a32c15a191b20e