If we have a visible cursor (caret), try to keep that stationary
This fixes at least one visible and annoying behaviour: when you zoom
in using the plus button, or the View>Zoom In menu entry, the caret
(text insertion cursor) location typically eventually will move
outside of the window. Then the window suddenly jumps to another
position so that the caret is again visible.

After this change, the zooming is much smoother and the caret location
stays visible all the time.

I could possibly have used the setZoomAround() but found it easier to
write the code incrementally actually understanding what I was
doing...

This does not fix the problem when the text insertion cursor is
already out of view when you start zooming in or out. In that case the
window still bluntly and unexpectedly jumps to make the location of
the text insertion cursor visible. In my opinion, it should not. I was
not able to figure out what causes that.

Change-Id: I594c4815f8fea9ebb43c0cb491d5167e603d724a