Update patch set 3
Patch Set 3:

When verifying this, note that you can be distracted by the unrelated bug that occasionally makes the caret show up in the wrong place after a zoom. So after each zoom step, if you see that the blinking text insertion cursor is not displayed where it actually should be, type a character and backspace over it right again (and note that the the caret is where it should be), and then zoom once more, etc.

Patch-set: 3