Make l10n of the messages from our JS code work in the mobile app
The l10n-for-node code does not work in the app. Instead of loading
the required JSON files at run-time, include all those translations
that seem complete enough as JavaScript code in the bundle.js. Use a
manaully curated list of translations, in a Perl script that generates
the JavaScript code in question.

Change-Id: I45d5cda477140ee63bb3fc8d9f1f4260bcdb97a7