Avoid TypeError and also possibly avoid the code working wrong
Found by code reading. Hopefully somebody who actually knows this code
can verify that this makes sense.

The for loop in the _addTiles function on each iteration assigns a
value to the tile variable conditionally, only if (coords.part ===
this._selectedPart) and (!this._tiles[key]). Otherwise tile stayed as
undefined (initially), or kept its previous value. After the large if
statement, tile.src was then in another if statement conditionally
assigned a value even if it still was undefined (which caused the
TypeError), or even if it actually had the value left over from the
previous loop iteration.

Make sure tile is set to undefined each loop iteration, and don't try
to use it if it is undefined. Also, if there *is* a this._tiles[key]
value, set tile to refer to that existing tile.

Change-Id: I0827814ab6fa6f701890060ac7c0ca72580dfc39