Update patch set 2
Patch Set 2:

I doubt many people want to use scrollbars for scrolling on a touch device anyway. At least I never do. All the scrollbars, or just the vertical one, is good for is to see how long the document is and how large portion of it you are viewing.

Patch-set: 2