Update patch set 2
Patch Set 2:

Hmm, now when I test after the revert, I notice that there is no nice physical inertia emulation when scrolling using a single-finger pan gesture on a mobile browser against normal online anyway. So that is not something this patch would have made go away. On a desktop browser, there is physics after two-finger scroll on touchpad (possibly the same holds for a wheel mouse, I don't have any such to test with).

Another thing that was already broken before this patch is that the scrollbar does not auto-hide automatically after single-finger panning or two-finger zooming on a mobile browser. On a desktop browser, it does auto-hide nicely,

Patch-set: 2