Update patch set 1
Patch Set 1: Verified-1

Still segfaults on Debian with system libnss, because libsqlite3.so.0 is a symlink hence isn't copied to the chroot.

    $ find /usr/lib/x86_64-linux-gnu -name 'libsqlite3.so.*' -printf '%y %p %l\n'
    f /usr/lib/x86_64-linux-gnu/libsqlite3.so.0.8.6
    l /usr/lib/x86_64-linux-gnu/libsqlite3.so.0 libsqlite3.so.0.8.6

Works when moving ‘usr/lib/x86_64-linux-gnu/libsqlite3.so*’ to `find … -type l`, though.  (That's also the what I did to make PDF export work on our instance.)  Or alternatively, calling ldd(1) not only on `xpdfimport`, but also on each of file matching usr/lib/x86_64-linux-gnu/nss/*.so.  (Neither option is very robust, unfortunately.)

Also, why hardcoding the architecture triplet here?  It's not hardcoded in other places (it's ‘lib64/ld-*’ not ‘lib64/ld-linux-x86-64.so.*’), so if ‘linux-x86-64’ isn't the only targeted triplet it's probably best to use globs here as well:

    find usr/lib/*/nss/*.so … -type f
    file usr/lib/*/libsqlite* … -type l

Patch-set: 1
Label: Verified=-1