Do pass Alt + non-printing keys through with the Alt modifier intact
Such commands are used for formatting functionality. (For instance,
Alt+Left or Alt+Right to shrink or grow a column in a spreadsheet, or
Alt+Enter to insert a paragraph just before a table in a text
document.)

All the uses of Alt in the meaning "alternate character input" on an
Apple keyboard are with keys for printing characters, like Alt+2 = @
on a Swedish/Finnish keyboard.

Change-Id: I6cfde57f540ac92c63420c8d7344102645050dd0