Update patch set 1
Patch Set 1: Code-Review-2

As discussed on IRC, the case for 'force autosave' is to make sure the latest modifications are saved before closing a document. The user might not even realize that the document is getting closed when they momentarily lose connection (or indeed when they accidentally close the browser or it crashes). In these cases an 'autosave' is triggered but it must be treated as good as the user's save, lest we lead the wopi host to think it can optimize it away as a regular checkpoint save, which it isn't if the document will get unloaded right after that autosave.

For the above reasons, autosave with force flag is treated like a user's save invocation; we can't afford to treat it as a checkpoint save, even if it isn't triggered by the user, it's on the user's behalf (assuming most users wouldn't want their changes to be lost by default).

As a workaround for cases to differentiate between autosaves because of document going away vs user invoked save, two client-side events can help. For user-issued save, intercept Action_Save and Action_SaveAs. For documents going away, intercept Close_Session (which you won't get when the browser crashes, connection is lost, etc.).

Patch-set: 1
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-2