Update patch set 2
Patch Set 2:

> Having said that, I agree that the closing of the document should be treated as a normal save forced by the user.

> What is the use case here, please?

This patch changes nothing for existing WOPI backends, for them closing of the document will still be treated as a normal save.

It only allows to filter putFile calls to see which calls were made by the user and which not.
And this call is definitely not made by the user.

In our use case, we want to make sure, that the document only gets saved, when the user wants to. If he only wants to play with it, it should not get saved.
We have other means to make sure that no data is lost.

Thus we need something to tell apart when the user wants to save the file and when the save operation was caused by cleaning up old documents.

Patch-set: 2