Make the 'closemobile' button (at left, looks like '<') work in the mobile app
We now have two places, the handler for the 'closemobile' button and
the handler for the 'closedocument' menu entry, where we have an if
statement that checks window.ThisIsAMobileApp and if so posts the "BYE"
message to the webkit message handler in the app. Possibly we could
move that conditional code into just one place, down into the 'close'
function in L.Socket instead?

Change-Id: Ib8c6adf621945b266fa4ea5f6d66b0022c2ca373