Call preventDefault() to improve selection handle dragging on touch devices
Now you can extend the selection also in the vertical direction by
dragging the handles. it doesn't cause the whole page to be dragged
any longer.

For some reason the handles don't follow the touching finger exactly,
but their distance from the original position where the touch drag
started is larger than that of the finger by a small factor, like 1.5
perhaps. Not sure whether that might be intentional, so that the
location of the handle isn't hidden by the finger. It might also be a
bug, perhaps related to Retina/HiDPI issues? (The same effect affects
the "teardrop" caret handle, too.)

Change-Id: Ic70e2b4090ea206ea6f15f9df667aad368afd641