Update patch set 1
Patch Set 1:

One problem with CI is that it's hard to say we should build against what core.git commit. Just latest core.git master? And if that's broken, live with it? (Similar to the situation when you base your good core change on a broken core.git master?)

Patch-set: 1