Update patch set 1
Patch Set 1:

Would it be a good idea to discuss this on the dev list? I see the benefits of consistent formatting, but at the same time there are costs of introducing it (think about backporting to release branches), especially when done for all files, not only for a selection of them (like done for core.git). It would be good to see what others see as benefits/costs as well.

Patch-set: 1