Update patch set 1
Patch Set 1:

> Would it be a good idea to discuss this on the dev list?
Don't think it's necessary to bikeshed on the dev list. This is a smallish repository and if the maintainers agree, I guess this can go in.

> I see the benefits of consistent formatting, but at the same time there are costs of introducing it (think about backporting to release branches), especially when done for all files, not only for a selection of them (like done for core.git).

I think the code already has quite consistent, good formatting. But as more and more people work on the code, it will become more and more unconsistent. So I am in favor of formatting the whole code, even if it makes backporting harder for a few months.

I think the benefits outweigh the costs as the codebase is still young and many lines stay untouched.

One thing I would probably change in the clang-format confi: Choose a different indentation style for namespaces so that we don't have to reformat the whole files like Util.hpp, SigUtil.hpp, etc.

Patch-set: 1