Update patch set 2
Patch Set 2: Code-Review-2

Please let's _not_ reformat in one huge patch.

If you want to get us to the consistent formatting, please do something like we do in the core.git - introduce a blacklist of files, start with the blacklist full, and then identify the files that are mostly well formatted, format them & remove from the blacklist (and from then on, they should remain formatted.  Of course, that all in combination with the appropriate git hook.

And then it is easy to discuss the remaining files: If they are actively worked on (and thus reformatting would harm potential backporting), or not really (and can be easily reformatted).

Thank you!

Patch-set: 2
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-2