leaflet: account for rounding error when scrolling toolbar
On some mobile browsers (Chrome) we get the maximum scroll
value that is 0.5 less than the real maximum, so we always
think we still have room to scroll more right. This
has the undesirable side-effect of keeping the >> button
visible, even when there is no way to scroll further right.

This allows for at most 1 pixel error.

Change-Id: Ie2e8902f1e2f4f358f210e9b0236422bc4f1da00