Update patch set 1
Patch Set 1:

Noel, not clear what your opinion about this change as such is; do you think it should NOT go in? Or are you just warning that like all the other places in loleaflet that are related to 'touchmove' events, deeper investigation might be needed in Chrome-based browsers after a year or so?

Patch-set: 1