Update patch set 1
Patch Set 1:

This patch sanitizes hyperlinks that can be opened via right click > Open hyperlink.

Although there are a ton of files in this commit, this is because of another reason: shrinkpack got a new update for modern Node versions (circa 2018 august).

If you ignore those, the changes are in loleaflet/README (talking about shrinkpack@next), loleaflet/package.json (where I build the dist version of a library, so that I may be able to use it as a global variable in the code) and in TileLayer.js (where the hyperlink checking actually takes place).

Patch-set: 1
Reviewer: Gerrit User 1002071 <1002071@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0