Update patch set 1
Patch Set 1:

I am sorry, I did not know there was a bugreport like this; we should have closed that one before you started the work on this :-(

We never want to deactivate the Save button, because the use might have reasons to use that even when the document is not modified.  That is for the cases when the user wants to save information that is not actually changing the 'document modified' status, but still is part of the file format.

Most notable example is the position of the view and of the cursor in a spreadsheet: When you move the cursor, the view and the cursor position changes, but the document modification status does not.  The users should have the possibility to save the spreadsheet when they just move the cursor.

We used to disable the Save button in the desktop version based on the 'document modified' status in the past, but changed that due to the reasons described above.

Sorry for that - technically you have achieved the Easy Hack, so if you need that for the GSoC, you can mention this as a completed one; but it is something we should not integrate.

Patch-set: 1
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0