Update patch set 3
Patch Set 3: -Code-Review

One missing piece here is the watermarking - which needs to be per-session; unfortunately this has quite some significant cost in the tilecache: we can no longer share the tilecache between views with different watermarks; pre master - that would mean building a set of unique watermark states, and putting an index of that into the key of the tilecache filenames I think - but that's a separate patch.

I wonder if this is actually necessary: now I read it, the wopifileinfo -looks- like it does what we need here.

Also - of course, we should have unit tests for this =)

Patch-set: 3
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=0