Update patch set 4
Patch Set 4:

> I think some URL dereffer work was ongoing with Szymon.

That's good to hear. We also have our own derefferer for the LibreOffice Online version that we are using. Considered making a patch to the community later, but we'll get in touch.

However this might still be worth merging until the derefferer is implemented?

P.S. It would be good if the dereferer server can be configured via a setting.

Patch-set: 4