loleaflet: constrained dragging for pie segment
The extra information for the graphic selection is now formatted
according to JSON syntax so that is easier to parse.

Dragging a pie segment now is constrained as on desktop and there is
also a preview of the pie segment while it is dragged.

On drag end a msg is sent to core with the new pie segment offset
from the pie center, which allows core to move the pie segment to
the correct position.

Change-Id: I7f76d21d622006fca5d0922c5932daec50b13836