Update patch set 1
Patch Set 1:

Not tested it, but I wonder how this works when we have a partial word already written in Online - and the IME is triggered to update/modify that word - but I assume we have a set of clear test cases for different languages like eg. Chinese / Japanese to work with here (?) I would imagine multi-ideogram words in Chinese might be fun. Did you lookup some tricky IME input sequences to test with ?

Patch-set: 1