Update patch set 1
Patch Set 1:

Interesting case, as I was investigating the IME should modify the partial word, only we receive 'compositionstart' if the IME appears and  subsequent 'compostionupdate' and 'compositionend'. It is generic I think, we are listeners of the data received from IME and they are sent to LO Core.

I have tested with Chinese in Gnome Shell desktop (of course the not semantic of words) only as logic UNICODE data that provides IME

Patch-set: 1