Update patch set 1
Patch Set 1: Code-Review+1

Fair enough. I guess the surprising thing about a well behaved IME is that it should get content and context from the cursor position in the document and effectively update / re-write the existing word that the cursor is on - though not sure if we have the API to do that today.

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=+1