Update patch set 1
Patch Set 1:

ThisIsAMobileApp in this case is not needed I think, I used that because the issue was visible only on desktop.

I think I can remove this condition in the next patch as I already pushed this commit state to master and it is always confusing when we use different commits on different branches.

Patch-set: 1
Reviewer: Gerrit User 1000664 <1000664@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0