Update patch set 1
Patch Set 1:

Iván: Thanks for this change!  Please is it a cleanup, or a fix of a problem?

Would be good to mention the reason / background in the commit message; one can read quickly enough from the patch directly what it removes & what it introduces instead :-) - but it's hard to deduct the reason of the change.

Patch-set: 1
Label: Verified=0