Revert "loleaflet: mobile: fix max-width screen size tablet"
This makes the iOS app return to how it looked last week.

If the reverted change really was intentional and carefully designed
for normal Online I don't know what to do. I assume there is no way to
have CSS that would be conditional on whether it is running in the iOS
app or in a normal browser?

This reverts commit 01f37fbbd33292de4d8f64fb753428669cbc34a7.

Change-Id: I04eead4b0200c4ca40af56718f92234fe896828e