loleaflet: Change setTimeout to setInterval in mCustomScrollbar
This *should* fix a bug where timers eat up 20% oc FPU time after
leafing loleaflet open for a long time.

The current nested setTimeout seems to work fine, but the hypothesis
here is that browsers that keep async stack traces are saving a
constantly-growing stack trace, so that the real slow-down doesn't
come from the JS code, but rather from the browser's development tools.

There seems to be no benefit from patching upstream, as mCustomScrollbar
seems to be unmaintained since 2016.

Change-Id: Ie496cd7dcedf443911655939064c54ebb76c2bbc