Update patch set 6
Patch Set 6: Verified+1 Code-Review+2

Mert: Thanks a lot for the work! :-)  From the code reading perspective it looks as I'd expect it, nice!

There is just a small nitpick, the const std::string filenameAnonym you've added in DocumentBroker::saveToStorageInternal is unused; might be good to clean it up in a follow up patch, or use appropriately for the logging.

Otherwise, I am not 100% sure how it works with the extension which is (correctly) omitted when performing the renamefile call; does the underlying core save with or without the extension?

Either way - let's get it in, thanks again!

Patch-set: 6
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2