Update patch set 7
Patch Set 7:

Ah i'll fix it with a following patch. About the extension. According to the documentation of wopi rename its not safe for user to enter extension since they can do it wrong. It's handled on the WOPI host according to the mimetype of the document. I made it like even if they enter the extension js part filters it out so just the name goes to the server.

Patch-set: 7