Update patch set 1
Patch Set 1:

> (4 comments)
 > 
 > Kaishu: Looks great in general, but see some nitpicks below...  Can
 > you please update the patch?

Actually I was testing the code then I forgot to replace the strings with the resources strings. I will update the patch :-)
Yes, We can show a progress bar for the loading of loleaflet.html but then there will be two progress bars, one for the loading of html file and other which is shown by loleaflet. Will it affect the UX.

Patch-set: 1
Label: Verified=0