Update patch set 1
Patch Set 1:

> Actually I was testing the code then I forgot to replace the
 > strings with the resources strings. I will update the patch :-)

Thanks! :-)

 > Yes, We can show a progress bar for the loading of loleaflet.html
 > but then there will be two progress bars, one for the loading of
 > html file and other which is shown by loleaflet. Will it affect the
 > UX.

OK, let's see how it goes - my concern is that we want to show the user that "something is happening" as soon as possible; if not immediately :-)  We could eg. have just one shared progressbar outside of the JavaScript, and for the loading of the document, just expose the values to the "outside" Java progressbar, and not paint anything in JavaScript.

But either way, let's keep this for later, you have more important stuff to deal with.

Patch-set: 1
Label: Code-Style=0